Added Login and Signup UI
